# New flash writing method with offset/memory/size alignment handling