--- a/cores/esp8266/Esp.h
+++ b/cores/esp8266/Esp.h
@@ -150,6 +150,7 @@ class EspClass {
 
         bool flashEraseSector(uint32_t sector);
         bool flashWrite(uint32_t offset, uint32_t *data, size_t size);
+        bool flashRead(uint32_t offset, uint8_t *data, size_t size);
         bool flashRead(uint32_t offset, uint32_t *data, size_t size);
 
         uint32_t getSketchSize();