--- a/cores/esp8266/Esp.h
+++ b/cores/esp8266/Esp.h
@@ -150,9 +150,10 @@ class EspClass {
         bool checkFlashCRC();
 
         bool flashEraseSector(uint32_t sector);
-        bool flashWrite(uint32_t offset, uint32_t *data, size_t size);
-        bool flashRead(uint32_t offset, uint8_t *data, size_t size);
+        bool flashWrite(uint32_t offset, const uint32_t *data, size_t size);
+        bool flashWrite(uint32_t offset, const uint8_t *data, size_t size);
         bool flashRead(uint32_t offset, uint32_t *data, size_t size);
+        bool flashRead(uint32_t offset, uint8_t *data, size_t size);
 
         uint32_t getSketchSize();
         String getSketchMD5();