Fixes issues raised in https://github.com/esp8266/Arduino/pull/7491#issuecomment-670452122