@yumkam 2020-08-07
I meant different issue: over-read of `data` (via `*ptr`), not over-read of flash.
(and I suspect `spi_flash_read`/`spi_flash_write` don't support `(size % 4) != 0`. Something like
[this](https://github.com/esp8266/Arduino/commit/b6bfe9907f5eae3f29817fe9358611270f4338b0)