@drzony 2020-08-07
flash_write_puya_buf size is checked against bytesNow, which is enough, since it's aligned to flash sector size:
```cpp
  // Good alternative for buffer size is: SPI_FLASH_SEC_SIZE (= 4k)
  // Always use a multiple of flash page size (256 bytes)
  #define PUYA_BUFFER_SIZE 256
```