@earlephilhower 2020-08-08
Thanks for the explanation.  I was going thru the original issue and thought the purpose was to not read past end-of-data.

This section of code has been very much trouble for us, so honestly what I'd need is a failing and passing valid test case on real XMC hardware with a couple folks' confirmations.  XMC just has some strange behavior which has caused something like 3 dot-releases so far, and stuff that looks like it should be fine often isn't when it meets that abomination of a flash chip.