@drzony 2020-08-08
I have PUYA chip in my Sonoff S20, so I'm verifying each time before creating a pull request.
A minimal test case for me is just a gzipped OTA upgrade (via Arduino OTA). Before my original patch from #7491 when gzip did not have 4 byte alignment the update failed, since the last `(size % 4)` bytes were not written at all. Unfortunately my first patch wrote additional `(4 - (size % 4))` bytes to the flash, that is what this one fixes. I'm still modifying `(size + 3) & ~3` in temporary buffer (which is the fix for gzipped OTA), but only reading/writing `size`.