@earlephilhower 2020-08-10
@devyte, if you're still relaxing, can you eyeball this?  Am I missing something?