@drzony 2020-08-10
@devyte 
AND is not out of bounds since the `flash_write_puya_buf` is 4 byte aligned and there is a check for `bytesNow > PUYA_BUFFER_SIZE`
@devyte @earlephilhower 
If the flash reads must be 4-byte aligned, then it should be fixed in `EspClass::flashWrite` and `EspClass::flashRead`, this means adding something similar to what is done currently for PUYA, for the generic case i.e.:
```cpp
if (!aligned) {
  readLast4BytesFromFlash();
  andReadWithData();
  writeLast4BytesToFlash();
}
```