@yumkam 2020-08-10
See [my suggested change linked above](https://github.com/esp8266/Arduino/commit/b6bfe9907f5eae3f29817fe9358611270f4338b0), it does exactly this (and also does not over-read `data`).