@drzony 2020-08-10
@yumkam It does not handle the generic case