@drzony 2020-08-10
My additional explanation about flash writes:
Extra write will not reduce life of flash:
1. SPI flashes can usually write in 1-byte increments, on SPI it's 4 probably due to speed (i.e. DIO/QIO instructions)
2. Flash life is usually measured in erase cycles
3. On SPI we can write to all bytes on page no matter the order, but only once before erase is needed
4. The additional write will not erase flash, so it will not wear it more
5. Example of such approach can be seen in Linux UBIFS driver, which keeps accounting of written "sub-pages" and fills the whole page before erasing it
6. The only problem here is when someone tries to write on unaligned offsets i.e. write 3 bytes at offset 0, then write 1 byte at offset 2. If `spi_flash_write` somehow aligns the write, which seems to be the case, then the second write will fail as the bytes 0-3 were already written.