@TD-er 2020-08-10
@drzony 
I don't agree with the statement about the writes.
The original PUYA chips for which the patch was made, is doing something in the background.
So what we're seeing is the data is exactly written to the flash as it is sent to it.
For flash memory, this is not what you would expect, as it should only change a 1 into a 0 and not the other way around.
But that's exactly what the PUYA chips do. If there is 0xAA stored in the flash and you write 0x55 to it, you would expect the cell to contain 0x00, but instead it holds 0x55.

So unless they have come up with some revolutionary new type of memory, they must perform an erase somewhere.
It could be the chip itself does buffer a page and does perform the erase/write cycle in the background, or it is using some page mapping.
But there must be extra erase/write cycles on those memory chips to get the result we're seeing.