@drzony 2020-08-10
@TD-er 
I can see the problem, but I don't have an idea on how to test whether read(4), write(4), read(4), write(4) is any different from read(8), write(8), because this is what seems to be the worry here.