@TD-er 2020-08-10
More like this:
write 3 bytes, start at 4N + 0  => read(4), write(4)
write 3 bytes, start at 4N + 3   => read(4), write(4), read(4), write(4)   (or read(8), write(8))