@drzony 2020-08-10
Ok, so I have confirmed that spi_flash_write self-aligns, this code:
```cpp
uint32_t flash_offset = 0xa0000;
uint32_t wr_buf1[2]   = {0x55555555, 0x55555555};
uint32_t wr_buf2[2]   = {0xAAAAAAAA, 0xAAAAAAAA};
uint32_t rd_buf1[2]   = {0x00000000, 0x00000000};

spi_flash_erase_sector(flash_offset / 0x1000);
spi_flash_write(flash_offset, wr_buf1, 3);
spi_flash_read(flash_offset, rd_buf1, 4);
DEBUG("Read from flash + 0 %08x\n", rd_buf1[0]);
flash_read(flash_offset + 4, rd_buf1, 4);
DEBUG("Read from flash + 4 %08x\n", rd_buf1[0]);
spi_flash_write(flash_offset + 4, wr_buf1, 4);
spi_flash_read(flash_offset + 8, rd_buf1, 4);
DEBUG("Read from flash + 8 %08x\n", rd_buf1[0]);
spi_flash_write(flash_offset + 8, wr_buf2, 4);
spi_flash_read(flash_offset + 4, rd_buf1, 4);
DEBUG("Read from flash + 4 %08x\n", rd_buf1[0]);
spi_flash_read(flash_offset + 8, rd_buf1, 4);
DEBUG("Read from flash + 8 %08x\n", rd_buf1[0]);
```
Produces (on esp12e with "normal" flash and on S20 with PUYA flash):
```txt
Read from flash + 0 55555555
Read from flash + 4 ffffffff
Read from flash + 8 ffffffff
Read from flash + 4 55555555
Read from flash + 8 aaaaaaaa
```
So always 4 bytes are written