@drzony 2020-08-10
Next test:
```cpp
uint32_t flash_offset = 0xa0000;
uint32_t wr_buf1[2]   = {0x55555555, 0x55555555};
uint32_t wr_buf2[2]   = {0xAAAAAAAA, 0xAAAAAAAA};
uint32_t rd_buf1[2]   = {0x00000000, 0x00000000};

spi_flash_erase_sector(flash_offset / 0x1000);
spi_flash_write(flash_offset, wr_buf1, 4);
spi_flash_write(flash_offset, wr_buf2, 4);
spi_flash_read(flash_offset, rd_buf1, 4);
DEBUG("Read from flash + 0 %08x\n", rd_buf1[0]);
```
Produces:
On "normal" flash:
```txt
Read from flash + 0 00000000
```
On PUYA flash:
```txt
Read from flash + 0 aaaa2aaa
```
So completely messed up.