@TD-er 2020-08-10
Hmm that `2` is strange in the PUYA output.
If it were only `a` HEX values, then I could understand it, but this `2` I can't explain.