@drzony 2020-08-10
And the last test for today:
```cpp
uint32_t flash_offset = 0xa0000;
uint32_t wr_buf1[2]   = {0xDEADBEEF, 0xDEADBEEF};
uint32_t rd_buf1[2] = {0x00000000, 0x00000000};
uint32_t rd_buf2[2] = {0x00000000, 0x00000000};
spi_flash_erase_sector(flash_offset / 0x1000);
spi_flash_read(flash_offset, rd_buf1, 4);
spi_flash_write(flash_offset, wr_buf1, 4);
spi_flash_read(flash_offset + 4, rd_buf1 + 1, 4);
spi_flash_write(flash_offset + 4, wr_buf1 + 1, 4);
spi_flash_read(flash_offset, rd_buf2, 8);
DEBUG("Read from flash + 0 %08x %08x %08x %08x\n", rd_buf1[0], rd_buf1[1], rd_buf2[0], rd_buf2[1]);
memset(rd_buf1, 0, 8);
memset(rd_buf2, 0, 8);
spi_flash_erase_sector(flash_offset / 0x1000);
spi_flash_read(flash_offset, rd_buf1, 8);
spi_flash_write(flash_offset, wr_buf1, 8);
spi_flash_read(flash_offset, rd_buf2, 8);
DEBUG("Read from flash + 0 %08x %08x %08x %08x\n", rd_buf1[0], rd_buf1[1], rd_buf2[0], rd_buf2[1]);
```
Produces
```txt
Read from flash + 0 ffffffff ffffffff deadbeef deadbeef
Read from flash + 0 ffffffff ffffffff deadbeef deadbeef
```
On both flashes, so if there is some erasing happening, then it's not affecting other data on flash.
As for the magic `2` in read, I also have no idea what is happening other than some weird buffering.