@drzony 2020-08-11
I've done more tests and it seems that whole `spi_flash_write_puya` is not needed, does anybody have a device that was used when the original patch was created and can confirm my findings?

My thoughts:
1. Writing multiple times to a single address on the flash, although sometimes possible, is not a good practice. One reason is that we can only change 1s to 0s, which would mean that we would have to verify if the data can be written (read flash, check if the only changes are 1s->0s, write or return error). The second reason is that not all flashes react well to this. Puya is one example, but I've also encountered some NAND flashes which had similar issues (even if they contained all 1s, the second write without erase would fail)
2. `spi_flash_write` always writes 4 bytes to the flash, so it triggers the previous point on unaligned write. I we do `spi_flash_write(4n + 0, data1, 3)` and `spi_flash_write(4n + 3, data2, 1)` my PR will work on "nice" flashes, but it will fail on "weird" flashes
3. The additional write in my PR will not cause any additional flash wear, since there is no erase.
4. As an additional "speedup" I can remove the additional `spi_flash_read` and initialize `tempData = 0xFFFFFFFF`, this will work on "nice" flashes as intended (I still think that writing multiple times is not a good way), it will fail on "weird" flashes regardless of whether we do `read` or set `FF`.
5. As for weird behavior of PUYA flashes, I think what we see can be explained by bitflips in hardware. The physical flash in those chips may need less power to flip to 1 (or is not keeping 0s that well), so that is why we see the change in value. This means that I do not agree with @TD-er that there must be some erase on the way, it can be explained by just poor persistence.